+++ DirectXTex/DirectXTexResize.cpp
@@ -811,7 +811,7 @@ namespace
 
         static_assert(TEX_FILTER_POINT == 0x100000, "TEX_FILTER_ flag values don't match TEX_FILTER_MASK");
 
-        unsigned long filter_select = filter & TEX_FILTER_MODE_MASK;
+        uint32_t filter_select = filter & TEX_FILTER_MODE_MASK;
         if (!filter_select)
         {
             // Default filter choice
